requirements.txt
streamlit==1.29.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
Pillow==10.1.0
pandas==2.1.3
accelerate==0.25.0
